Clean up styles
